requirements.txt
python-telegram-bot
pytest
pytest-asyncio
pdf2image
pydantic
pydantic-settings